project_wallet_Rest/wallet/requirements.txt
Django>=4.1.3
djangorestframework>=3.15.2
psycopg2-binary